requirements.txt
cycler==0.11.0
fonttools==4.28.5
future==0.18.2
iso8601==1.0.2
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
